Refactor: Split RecentTransactions component
Refactored the RecentTransactions component into smaller, more manageable files to improve code organization and readability. No functionality was changed. Unused imports and files were removed.
